Added `--fast` flag to speed up `Why` execution when testing in C#.
Since finding the answer to `Why` is not easy - it requires the use of some additional computational resources. When testing, it is enough to use the fast version, since `Why` is computed in many other languages at once, which in case of overloading may lead to an unintended collapse of the universe.